use NSInteger instead of raw type

+++ JBoss Admin/JBAAddDeploymentViewController.m
@@ -156,8 +156,8 @@
     
 #pragma mark - Table Delegate Methods
 - (void)tableView:(UITableView *)tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath {
-    int newRow = [indexPath row];
-    int oldRow = (_lastIndexPath != nil) ? [_lastIndexPath row] : -1;
+    NSInteger newRow = [indexPath row];
+    NSInteger oldRow = (_lastIndexPath != nil) ? [_lastIndexPath row] : -1;
     
     if (newRow != oldRow) {
         UITableViewCell *newCell = [tableView cellForRowAtIndexPath:indexPath];
